Refactored ReservationConfirmationEmailService to use itnerface and implementation
